[Rename] dto to model
