+++ src/main/webapp/WEB-INF/views/question/question-write.jsp
@@ -1,5 +1,5 @@
 <%@page import="java.util.ArrayList"%>
-<%@page import="com.saeromteo.app.dto.question.QuestionDTO.QuestionResponse"%>
+<%@page import="com.saeromteo.app.model.question.QuestionDTO.QuestionResponse"%>
 <%@page import="java.util.List"%>
 <%@ page language="java" contentType="text/html; charset=UTF-8"
 	pageEncoding="UTF-8"%>
